+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/taufiqgiri/fancy-todo#readme",
   "dependencies": {
+    "axios": "^0.21.0",
     "bcryptjs": "^2.4.3",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
